Lexer: improve debugging output & ensure the lexer isn't run multiple times at once
Another unexpected use for JS Symbols!

This way, you can transparently give up on Lexing something and start 
again, and we won't care.

But if you try and continue a previous lexing session on the same 
instance, we can efficiently and reliably detect it by checking the 
symbol after we return from every yield.